Major refactor for smt & mpt logic
- Because we're going to be dealing with feature flags for SMT stuff,
  it's going to be very worth while I think splitting the logic into
  seperate modules.
- This commit does just that.
