+++ trace_decoder/src/compact/mod.rs
@@ -1,4 +1,7 @@
-pub mod compact_prestate_processing;
+pub mod compact_debug_tools;
+pub mod compact_mpt_processing;
+mod compact_processing_common;
+pub mod compact_smt_processing;
 pub mod compact_to_mpt_trie;
 pub mod compact_to_smt_trie;
